[chore] viewModel네이밍 수정
